Change directory layout to be a standard Maven directory layout
## What?
This PR changes the layout of the project so that it adheres to the maven directory layout, which is recommended by the Bazel authors.
Also change the java package name to `com.bazel_diff` instead of `com.bazel-diff` because hyphens are not allowed in java package names.

## Why?
I wanted to experiment with the project so I opened it in IntelliJ with the Bazel plugin and nothing worked.
By using this folder structure and adding package names to all files it works very well.