Fix computing external repo fine-grained hashes
The previous implementation ignores any change of source files in
external repos. That implementation works for maven_install because
maven_install does not download any jars into that particular repo. But
it does not work for cases when the external repo comes with
sources/jars in itself.

We have such use cases internally in Pinterest though I am not aware of
any such use cases in common open source world.